refactor: 패키지 구조 변경 및 체스판에 말을 추가하는 테스트케이스 추가
[- chess, pieces 패키지 추가
- 체스판에 말만 추가할 수 있음을 검증하는 테스트케이스 추가]
