+++ src/test/java/com/wootecam/chess/BoardTest.java
@@ -1,14 +1,17 @@
-package com.wootecam;
+package com.wootecam.chess;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.wootecam.chess.pieces.Pawn;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -19,6 +22,18 @@ import org.junit.jupiter.params.provider.ValueSource;
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @DisplayName("체스판 테스트")
 class BoardTest {
+
+    @Nested
+    class 체스판에_말을_추가할_수_있다 {
+
+        @Test
+        void 말을_추가할_수_있다() {
+            var board = new Board();
+            var pawn = new Pawn();
+
+            assertThatNoException().isThrownBy(() -> board.add(pawn));
+        }
+    }
 
     @Nested
     class 체스판에_추가된_말의_개수를_조회할_수_있다 {
@@ -35,7 +50,6 @@ class BoardTest {
         @ParameterizedTest
         void 말의_개수를_조회할_수_있다(List<Pawn> pawnList) {
             var board = new Board();
-
             for (Pawn p : pawnList) {
                 board.add(p);
             }
@@ -59,7 +73,6 @@ class BoardTest {
         @ParameterizedTest
         void 말을_조회할_수_있다(List<Pawn> pawnList) {
             var board = new Board();
-
             for (Pawn p : pawnList) {
                 board.add(p);
             }
